requirements.txt
fastapi
uvicorn[standard]
pymupdf
pandas
python-multipart
pathway
google-generativeai
#sse_starlette